fix dojo state issues when switching networks
