+++ ui/src/dojo/hooks/useSdkQueries.ts
@@ -18,7 +18,7 @@ export const useGetTokensQuery = (namespace: string) => {
         .withClause(KeysClause([getModelsMapping(namespace).Token], []).build())
         .withEntityModels([getModelsMapping(namespace).Token])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
 
   const { entities, isLoading, refetch } = useSdkGetEntities({
@@ -46,7 +46,7 @@ export const useGetMetricsQuery = (namespace: string) => {
           getModelsMapping(namespace).PrizeMetrics,
         ])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
   const { entities, isLoading, refetch } = useSdkGetEntities({
     query,
@@ -119,7 +119,7 @@ export const useGetTournamentQuery = (
         ])
         .includeHashedKeys()
         .withLimit(10000),
-    [tournamentId]
+    [tournamentId, namespace]
   );
 
   const { entities, isLoading, refetch } = useSdkGetEntities({
@@ -179,7 +179,7 @@ export const useGetRegistrationsForTournamentInTokenListQuery = ({
         )
         .withEntityModels([getModelsMapping(namespace).Registration])
         .includeHashedKeys(),
-    [tournamentId, tokenIds, limit, offset]
+    [tournamentId, tokenIds, limit, offset, namespace]
   );
 
   const { entities, isLoading, refetch } = useSdkGetEntities({
@@ -323,7 +323,7 @@ export const useSubscribeTournamentQuery = (
           getModelsMapping(namespace).Prize,
         ])
         .includeHashedKeys(),
-    [tournamentId]
+    [tournamentId, namespace]
   );
 
   const { entities, isSubscribed } = useSdkSubscribeEntities({
@@ -339,7 +339,7 @@ export const useSubscribePrizesQuery = (namespace: string) => {
         .withClause(KeysClause([getModelsMapping(namespace).Prize], []).build())
         .withEntityModels([getModelsMapping(namespace).Prize])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
 
   const { entities, isSubscribed } = useSdkSubscribeEntities({
@@ -360,7 +360,7 @@ export const useSubscribeTournamentsQuery = (namespace: string) => {
         )
         .withEntityModels([getModelsMapping(namespace).Tournament])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
 
   const { entities, isSubscribed } = useSdkSubscribeEntities({
@@ -378,7 +378,7 @@ export const useSubscribeTokensQuery = (namespace: string) => {
         )
         .withEntityModels([getModelsMapping(namespace).Token])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
 
   const { entities, isSubscribed } = useSdkSubscribeEntities({
@@ -405,7 +405,7 @@ export const useSubscribeMetricsQuery = (namespace: string) => {
           getModelsMapping(namespace).PrizeMetrics,
         ])
         .includeHashedKeys(),
-    []
+    [namespace]
   );
 
   const { entities, isSubscribed } = useSdkSubscribeEntities({
